#52: Promote simple-funcs dependency

+++ pancake-platform/src/main/java/dev/kkorolyov/pancake/platform/registry/internal/ActionResourceReaderFactory.java
@@ -30,7 +30,7 @@ public final class ActionResourceReaderFactory implements ResourceReaderFactory.
 
 	private static final Pattern MULTI_STAGE_SPLIT_PATTERN = Pattern.compile(",\\s?(?![^\\[]*])");
 	private static final Pattern MULTI_STAGE_PATTERN = Pattern.compile("\\{.*(,.*){2}}");
-	private static final long MULTI_STAGE_HOLD_THRESHOLD = (long) (Float.parseFloat(Config.config().get("holdThreshold")) * 1e9);
+	private static final long MULTI_STAGE_HOLD_THRESHOLD = (long) (Double.parseDouble(Config.config().get("holdThreshold")) * 1e9);
 
 	private static final Pattern KEY_PATTERN = Pattern.compile("\\[[_a-zA-Z]+(,\\s*[_a-zA-Z]+)*]\\s*=\\s*.*");
 	private static final Pattern KEY_SPLIT_PATTERN = Pattern.compile("\\s*=\\s*");
